fix mismatch between decoded MPEG IMX height and height assumed when allocating buffer
FFmpeg (March 2011) includes the 32 or 26 lines of VBI in the decoded MPEG
IMX frame. The buffer allocated for the decoded frame did not take this
into account and a segmentation fault resulted when trying to copy too many
lines.

The fix is to use the AAF FlavourParameters_t dimensions rather than the
FFmpeg decoded picture dimensions when copying the data. It is asserted
that the FFmpeg decoded picture dimensions are always <= and
therefore that enough space has been allocated.

The output picture defaults to the top-left corner of the decoded picture.
For MPEG IMX the picture is shifted down (if the decoded picture
included the VBI lines, which it might not in older or future FFmpeg versions).