Fix IAAFTypeDefStream3::Read() and IAAFPlainStreamData2::Read() to allow reading past the stream end.
Debug-only assertions in ImplAAFStreamPropertyValue::ReadFiltered() and ImplAAFStreamPropertyValue::Read() do not allow the caller to read beyond the stream boundaries with a debug build. This is inconsistent with support for reading past the stream end in other class' read methods (see, for example, AAFRESULT_END_OF_DATA error support and lack of similar assertions in other methods).

These changes extend full support for reading past the stream end to ReadFiltered() and Read():
    - Add AAFTypeDefStream test for reading past the end of the stream.
    - Change ImplAAFStreamPropertyValue::ReadFiltered() and ImplAAFStreamPropertyValue::Read() to remove assertions that expect the caller to read within the stream boundaries.