Ignore non-standard PrimaryMob property definition when opening files that have one. (Fixes access violation during save when modifying existing files.)
The change deals with files that define the PrimaryMob property using non-standard dynamic local tag (PID). A dynamic PID conflicts with the standard static PID (0x3b08) in the AAF SDK built-in property definition, which results in uninitialized property definition pointer in OMProperty, triggers access violation during save and makes it impossible to modify the file with AAF SDK.

The fix is to replace the PrimaryMob property definition in the file's dictionary with the definition from the SDK's built-in dictionary. The behavior with files that have PrimaryMob instances using the invalid property definition is unknown.