Do no re-use dynamic property ID already assigned in existing file when registering definition of different optional property.
Bug in opening MXF files created outside the SDK: Property definition
registered by the SDK when reading a metadata object from MXF file may
get assigned dynamic property ID (PID) that conflicts with the same
property ID used by different property in the file. The result is 2
property definitions with the same PID in the dictionary leading to an
assertion violation in OMPropertySet::put(OMProperty*) trying to
instantiate a second of the two.

The problem is in ImplAAFDictionary::GenerateOmPid() that fails to
account for PIDs assigned in the file's MXF Primer Pack.

The fix is to:
  - Update ImplAAFBuiltinClasses class to allow access to property IDs
  (PIDs) assigned to property definitions in an existing MXF file.
  - Account for PIDs already assigned in an existing file when
    initializing the list of allocated PID segments in
    ImplAAFDictionary::GenerateOmPid().